% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_char_time.R
\name{as_char_time}
\alias{as_char_time}
\title{as_char_time}
\usage{
as_char_time(x)
}
\arguments{
\item{x}{time vector eg: integer, times, or ITime classes}
}
\description{
convert time values to character strings
}
\details{
takes a vector of times, and converts to character vector
}
