% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_Date.R
\name{as_Date}
\alias{as_Date}
\title{as_Date}
\usage{
as_Date(x)
}
\arguments{
\item{x}{character vector of date values, formatted as YYYY-mm-dd. the hyphens can be anything}
}
\description{
convert character time values to Date format
}
\details{
takes a character vector of dates, and converts to Date class vector
}
