% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_IDate.R
\name{as_IDate}
\alias{as_IDate}
\title{as_IDate}
\usage{
as_IDate(x)
}
\arguments{
\item{x}{character vector of date values, formatted as YYYY-mm-dd. the hyphens can be anything}
}
\description{
convert character time values to IDate format
}
\details{
takes a character vector of dates, and converts to IDate class vector
}
