% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ITime.R
\name{as_ITime}
\alias{as_ITime}
\title{as_ITime}
\usage{
as_ITime(x)
}
\arguments{
\item{x}{character vector of time values}
}
\description{
convert character time values to ITime format
}
\details{
takes a character vector of times, and converts to ITime class vector
}
