% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_times.R
\name{as_times}
\alias{as_times}
\title{as_times}
\usage{
as_times(x)
}
\arguments{
\item{x}{character vector of time values}
}
\description{
convert character time values to times format
}
\details{
takes a character vector of times, and converts to times class vector
}
