% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asCharTime}
\alias{asCharTime}
\title{asCharTime}
\usage{
asCharTime(x)
}
\arguments{
\item{x}{integer vector eg: `ITime` class}
}
\description{
convert integer time values to character time strings
}
\keyword{time}
