% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_rcsv.R
\name{classes_rcsv}
\alias{classes_rcsv}
\title{classes_rcsv}
\usage{
classes_rcsv(file)
}
\arguments{
\item{file}{file path from which the rscv will be read}
}
\value{
named character vector of column classes
}
\description{
extract column classes from an rcsv file
}
\details{
extract column classes from an rcsv file without loading
the entire file into memory
}
