% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_rcsv.R
\name{dim_rcsv}
\alias{dim_rcsv}
\title{dim_rcsv}
\usage{
dim_rcsv(file)
}
\arguments{
\item{file}{file path from which the rscv will be read}
}
\value{
named integer vector of length 2. "rows" and "cols"
}
\description{
retrieve the dimensions of a table stored in an rcsv file
}
\details{
retrieve dimensions of a table stored in an rcsv file,
without loading the entire dataset into memory
}
