% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_rcsv.R
\name{glimpse_rcsv}
\alias{glimpse_rcsv}
\title{glimpse_rcsv}
\usage{
glimpse_rcsv(file, width = 80)
}
\arguments{
\item{file}{file path to which the rscv will be written}

\item{width}{integer, approximate character length of each line
of the output display}
}
\description{
display a short summary of the contents of an rcsv
file without loading the entire dataset, similar to
dplyr::glimpse for a data frame in memory
}
\details{
read in and print the `head` of an rcsv file. For
viewing data, not to be used for data input
}
