% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_rcsv.R
\name{names_rcsv}
\alias{names_rcsv}
\title{names_rcsv}
\usage{
names_rcsv(file)
}
\arguments{
\item{file}{file path from which the rscv will be read}
}
\description{
extract column names from an rcsv file
}
\details{
extract column names from an rcsv file without loading
the entire file into memory
}
