% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_add.R
\name{notes_add}
\alias{notes_add}
\title{notes_add}
\usage{
notes_add(x, notes)
}
\arguments{
\item{x}{object to which the note will be added}

\item{notes}{character vector of notes to add}
}
\description{
add to an existing object note
}
\details{
add a note to an object, retaining any existing notes.
Specifically intended for use with objects being stored as `rcsv`
}
