% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_rcsv.R
\name{notes_rcsv}
\alias{notes_rcsv}
\title{notes_rcsv}
\usage{
notes_rcsv(x)
}
\arguments{
\item{x}{object name or file path from which the rscv will be read}
}
\description{
extract notes from an rcsv file or object
}
\details{
display dataset notes, extracting from an rcsv file if necessary,
without loading the entire file into memory
}
