% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_replace.R
\name{notes_replace}
\alias{notes_replace}
\title{notes_replace}
\usage{
notes_replace(x, notes)
}
\arguments{
\item{x}{object to which the note will be added}

\item{notes}{character vector of notes}
}
\description{
add a note to an object, deleting any existing notes
}
\details{
add a note to an object, deleting any existing notes.
Specifically intended for use with objects being stored as `rcsv`
}
