% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toTimeDay}
\alias{toTimeDay}
\title{toTimeDay}
\usage{
toTimeDay(x)
}
\arguments{
\item{x}{character vector, input times}
}
\description{
convert character times to fraction of the day
}
\keyword{time}
