% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{toTimeSec}
\alias{toTimeSec}
\title{toTimeSec}
\usage{
toTimeSec(x)
}
\arguments{
\item{x}{character vector, input times}
}
\description{
convert character times to integer seconds since midnight
}
\keyword{time}
