% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_offset.R
\name{tz_offset}
\alias{tz_offset}
\title{tz_offset}
\usage{
tz_offset(x)
}
\arguments{
\item{x}{POSIXct value}
}
\description{
find the offset from UTC applied to a POSIXct object
}
\details{
find the offset from UTC applied to a POSIXct object
}
