% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_rcsv.R
\name{head_rcsv}
\alias{head_rcsv}
\title{head_rcsv}
\usage{
head_rcsv(file, n = 6, echo.notes = TRUE)
}
\arguments{
\item{file}{file path from which the rscv will be read}

\item{n}{integer, number of rows to read}

\item{echo.notes}{print notes to the console on import?}
}
\description{
read only the first few lines of an rscv file
}
\details{
read in the `head` of an rcsv file, without loading the
entire file
}
