% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_rcsv.R
\name{notes_rcsv}
\alias{notes_rcsv}
\title{notes_rcsv}
\usage{
notes_rcsv(file)
}
\arguments{
\item{file}{file path from which the rscv will be read}
}
\description{
extract notes from an rcsv file
}
\details{
extract notes from an rcsv file without loading
the entire file into memory
}
